add api calls to profile .js not auth.js
